+++ src/view/admin/SoulsSetting.ts
@@ -51,8 +51,12 @@ export default class SoulsSetting extends View {
         for (const bio of bios) {
             new SoulDisplay(bio, AdminLayout.current.bio.color, el("a.unlink", el("i.fa-light.fa-link-slash"), {
                 click: async () => {
-                    await SoulinkContract.breakLink(await SoulinkContract.getTokenId(bio.id!));
-                    new Alert("The transaction has been registered. Please wait until it is finished.");
+                    if (AdminLayout.current.address !== await Wallet.loadAddress()) {
+                        new Alert(`Whoa! Looks like you’ve signed in with another account. Please sign back in with ${AdminLayout.current.address} to continue.`);
+                    } else {
+                        await SoulinkContract.breakLink(await SoulinkContract.getTokenId(bio.id!));
+                        new Alert("The transaction has been registered. Please wait until it is finished.");
+                    }
                 },
             })).appendTo(this.linkedContainer!);
         }
@@ -75,42 +79,50 @@ export default class SoulsSetting extends View {
                 const requestDisplay = new SoulDisplay(request.bio ?? { id: request.requester, links: [] }, AdminLayout.current.bio.color,
                     el("a", el("i.fa-light.fa-check"), {
                         click: async () => {
-                            const deadline = Math.floor(Date.now() / 1000) + 315360000; // +10년
-                            const signature = await Wallet.signTypedData(AdminLayout.current.address, "Soulink", "1", SoulinkContract.address, "RequestLink", [
-                                { name: "to", type: "address" },
-                                { name: "deadline", type: "uint256" },
-                            ], {
-                                to: request.requester,
-                                deadline,
-                            });
-                            await fetch(`${Config.apiURI}/accept`, {
-                                method: "POST",
-                                body: JSON.stringify({
-                                    requester: request.requester,
-                                    target: AdminLayout.current.address,
-                                    signature,
+                            if (AdminLayout.current.address !== await Wallet.loadAddress()) {
+                                new Alert(`Whoa! Looks like you’ve signed in with another account. Please sign back in with ${AdminLayout.current.address} to continue.`);
+                            } else {
+                                const deadline = Math.floor(Date.now() / 1000) + 315360000; // +10년
+                                const signature = await Wallet.signTypedData(AdminLayout.current.address, "Soulink", "1", SoulinkContract.address, "RequestLink", [
+                                    { name: "to", type: "address" },
+                                    { name: "deadline", type: "uint256" },
+                                ], {
+                                    to: request.requester,
                                     deadline,
-                                }),
-                            });
-                            requestDisplay.delete();
-                            new Alert("Soulink accepted.");
+                                });
+                                await fetch(`${Config.apiURI}/accept`, {
+                                    method: "POST",
+                                    body: JSON.stringify({
+                                        requester: request.requester,
+                                        target: AdminLayout.current.address,
+                                        signature,
+                                        deadline,
+                                    }),
+                                });
+                                requestDisplay.delete();
+                                new Alert("Soulink accepted.");
+                            }
                         },
                     }),
                     el("a", el("i.fa-light.fa-xmark"), {
                         click: async () => {
-                            const signedMessage = await Wallet.signMessage("Ignore the request.");
-                            const loading = new Loading("Ignoring...").appendTo(this.container!);
-                            await fetch(`${Config.apiURI}/ignore`, {
-                                method: "POST",
-                                body: JSON.stringify({
-                                    signedMessage,
-                                    requester: request.requester,
-                                    target: AdminLayout.current.address,
-                                }),
-                            });
-                            loading.delete();
-                            requestDisplay.delete();
-                            new Alert("Ignored.");
+                            if (AdminLayout.current.address !== await Wallet.loadAddress()) {
+                                new Alert(`Whoa! Looks like you’ve signed in with another account. Please sign back in with ${AdminLayout.current.address} to continue.`);
+                            } else {
+                                const signedMessage = await Wallet.signMessage("Ignore the request.");
+                                const loading = new Loading("Ignoring...").appendTo(this.container!);
+                                await fetch(`${Config.apiURI}/ignore`, {
+                                    method: "POST",
+                                    body: JSON.stringify({
+                                        signedMessage,
+                                        requester: request.requester,
+                                        target: AdminLayout.current.address,
+                                    }),
+                                });
+                                loading.delete();
+                                requestDisplay.delete();
+                                new Alert("Ignored.");
+                            }
                         },
                     }),
                 ).appendTo(this.toAcceptContainer!);
@@ -123,15 +135,19 @@ export default class SoulsSetting extends View {
             ) {
                 const soulDisplay = new SoulDisplay(request.bio ?? { id: request.target, links: [] }, AdminLayout.current.bio.color, el("a.link", el("i.fa-light.fa-link"), {
                     click: async () => {
-                        await SoulinkContract.setLink(await SoulinkContract.getTokenId(request.target), [
-                            request.signature,
-                            request.accept!.signature,
-                        ], [
-                            request.deadline,
-                            request.accept!.deadline,
-                        ]);
-                        soulDisplay.delete();
-                        new Alert("The transaction has been registered. Please wait until it is finished.");
+                        if (AdminLayout.current.address !== await Wallet.loadAddress()) {
+                            new Alert(`Whoa! Looks like you’ve signed in with another account. Please sign back in with ${AdminLayout.current.address} to continue.`);
+                        } else {
+                            await SoulinkContract.setLink(await SoulinkContract.getTokenId(request.target), [
+                                request.signature,
+                                request.accept!.signature,
+                            ], [
+                                request.deadline,
+                                request.accept!.deadline,
+                            ]);
+                            soulDisplay.delete();
+                            new Alert("The transaction has been registered. Please wait until it is finished.");
+                        }
                     },
                 })).appendTo(this.toLinkContainer!);
             }
